requirements.txt
numpy
pandas
matplotlib
scikit-learn
scipy
PyYAML
torch
covmetrics